Changes in response to comments. Javadocs changed. Avro to Structured Tests in separate files
